+++ documents/guide/history.md
@@ -1,8 +1,10 @@
 ## 更新日志
 
-> 2025-01-16 版本号：`0.0.12`
+> 2025-01-16 版本号：`0.0.13`
 1. `addButton`,`editButton`,`deleteButton`三个按钮默认添加`e.stopPropagation()`阻止冒泡
 2. 完善更多的typescript类型
+3. ipage添加`reloadForm`参数，用于打开表单时候重载IForm
+4. 减少selectMore减少出现Uncaught ResizeObserver loop completed with undelivered notifications.的情况，目前只有在弹窗中打开后直接关闭弹窗会有
 
 > 2025-01-15 版本号：`0.0.10`
 1. IForm的渲染单元isCol默认打开，默认span=12，确保和旧版本逻辑一致